Feat: Add an authentication system with JWT and CORS config change
Implemented JWT-based authentication including login, logout, token refresh, and user registration functionalities. Changing the CORS configuration. Introduced repositories, services, and configuration classes to support user management and role-based security.

+++ src/test/java/com/edusyspro/api/repository/ScoreRepositoryTest.java
@@ -51,8 +51,8 @@ public class ScoreRepositoryTest {
         scoreRepository.saveAll(scores);
     }
 
-    public float getMark() {
+    public Double getMark() {
         final Random random = new Random();
-        return random.nextInt(21);
+        return random.nextDouble(21);
     }
 }
